+++ last-order-api/src/main/java/com/flab/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.flab.user.service;
 
 import com.flab.user.domain.User;
-import com.flab.user.dto.LoginRequest;
+import com.flab.user.dto.request.LoginRequest;
 
 public interface UserService {
     User signUp(User user);
