+++ last-order-api/src/main/java/com/flab/user/service/impl/UserServiceImpl.java
@@ -2,7 +2,7 @@ package com.flab.user.service.impl;
 
 
 import com.flab.user.domain.User;
-import com.flab.user.dto.LoginRequest;
+import com.flab.user.dto.request.LoginRequest;
 import com.flab.user.exception.EmailAlreadyExistException;
 import com.flab.user.exception.EmailNotExistException;
 import com.flab.user.exception.PasswordNotMatchException;
@@ -11,8 +11,6 @@ import com.flab.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.mindrot.jbcrypt.BCrypt;
 import org.springframework.stereotype.Service;
-
-import static com.flab.common.ErrorMessage.*;
 
 @Service
 @RequiredArgsConstructor
@@ -27,7 +25,6 @@ public class UserServiceImpl implements UserService {
         String encrypt = BCrypt.hashpw(user.getPassword(), BCrypt.gensalt());
         user.setPassword(encrypt);
         userRepository.save(user);
-
         return user;
     }
